app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
awscli==1.27.104
botocore==1.29.104
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
docutils==0.16
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
mangum==0.17.0
multidict==6.0.4
openai==0.27.2
pyasn1==0.4.8
pydantic==1.10.7
pyinotify==0.9.6
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
repoze.lru==0.7
requests==2.28.2
requestsexceptions==1.4.0
rsa==4.7.2
s3transfer==0.6.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
statsd==4.0.1
Tempita==0.5.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.4
yappi==1.4.0
yarl==1.8.2
